+++ pakyow-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `fix` **Allow unmounted applications to be looked up but not receive requests.**
+
+    *Related links:*
+    - [Pull Request #500][pr-500]
+
   * `fix` **Correctly duplicate validator/verifier state.**
 
     *Related links:*
@@ -605,6 +610,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-500]: https://github.com/pakyow/pakyow/pull/500
 [pr-499]: https://github.com/pakyow/pakyow/pull/499
 [pr-497]: https://github.com/pakyow/pakyow/pull/497
 [pr-496]: https://github.com/pakyow/pakyow/pull/496